Lists of integration time and telescope location
This only iterates on the three FITS files stored on github, this will
be updated to iterate over all FITS files on the nisus soon. The
integration times calculated are deviations from the expected 900
second integration time. I have not been able to access the nisus yet
so I am not sure what the telescope ids are for all telescopes. Right
now it checks to see if the telescope location matches for the one in
Chile and assigns this as the location.